Move eviction stuff to eviction/ directory
